[core] cosmetics (short copyright notices)
