+++ src/xci/core/error.h
@@ -1,4 +1,8 @@
-// error.h - created by Radek Brich on 2018-09-23
+// error.h created on 2018-09-23 as part of xcikit project
+// https://github.com/rbrich/xcikit
+//
+// Copyright 2018 Radek Brich
+// Licensed under the Apache License, Version 2.0 (see LICENSE file)
 
 #ifndef XCI_CORE_ERROR_H
 #define XCI_CORE_ERROR_H
