+++ src/xci/core/file.h
@@ -1,17 +1,8 @@
-// file.h created on 2018-03-29, part of XCI toolkit
+// file.h created on 2018-03-29 as part of xcikit project
+// https://github.com/rbrich/xcikit
+//
 // Copyright 2018 Radek Brich
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// Licensed under the Apache License, Version 2.0 (see LICENSE file)
 
 #ifndef XCI_CORE_FILE_H
 #define XCI_CORE_FILE_H
